Use Random() to generate seeded numbers
The proper way to create random numbers from a seed in Python is to
instantiate a `Random()` object with the desired seed, which properly
decouples the new PRNG from any calls to `random.random()`.

Also: replacing calls to `random()` with a constant function is a bad
habit to get into.